requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.8
django-autoslug==1.9.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
PyYAML==6.0.3
psycopg==3.3.2
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
gunicorn==21.2.0
dotenv~=0.9.9
python-dotenv~=1.2.1
